fix delete user method

+++ app/controllers/users_controller.rb
@@ -1,5 +1,5 @@
 class UsersController < ApplicationController
-  before_filter :authorize, only: [:edit, :update, :show, :delete]
+  before_filter :authorize, only: [:edit, :update, :show]
 
   def index
     @users = User.all
@@ -30,20 +30,26 @@ class UsersController < ApplicationController
 
   def update
     @user = User.find(params[:id])
-    if @user.update(user_params)
-      @user.update_attribute(:avatar, params[:user][:avatar])
-      flash[:notice] = "User updated."
-      redirect_to session_user_path(session, @user)
-    else
+    if current_user.id == @user.id
+      if @user.update(user_params)
+        @user.update_attribute(:avatar, params[:user][:avatar])
+        flash[:notice] = "User updated."
+        redirect_to session_user_path(session, @user)
+      else
       render 'edit'
+      end
     end
   end
 
   def destroy
     @user = User.find(params[:id])
-    if @user.destroy
-      flash[:notice] = "The user was deleted."
-      redirect_to root_path
+    if current_user.id == @user.id
+      if @user.destroy
+        current_user.destroy
+        session[:user_id] = nil
+        flash[:notice] = "The user was deleted."
+        redirect_to root_path
+      end
     end
   end
